backend/requirements.txt
fastapi[all]
langchain
langchain-community
langchain-openai
langchain-weaviate
openai
fastapi
uvicorn
python-dotenv
python-multipart
tiktoken
